finalizado

+++ 5periodo/1bim/trabalhonodejs/public/usuario.js
@@ -3,7 +3,7 @@ const modalcadastro = new bootstrap.Modal(document.getElementById('modalcadastro
 var idveiculoatual;
 
 function alterar(id) {
-  fetch("http://127.0.0.1:3333/veiculos/" + id)
+  fetch("http://127.0.0.1:3000/veiculos/" + id)
     .then(resp => resp.json())
     .then(dados => {
       idveiculoatual = id; 
@@ -18,11 +18,26 @@ function alterar(id) {
 }
 
 function excluir(id) {
-  fetch("http://127.0.0.1:3333/veiculos/" + id, {  
-    method: "DELETE",
-  })
-  .then(() => listar())
-  .catch(err => console.error("Erro ao excluir veículo:", err));
+  Swal.fire({
+    title: 'Tem certeza?',
+    text: "Essa ação não pode ser desfeita!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#c8102e',
+    cancelButtonColor: '#000',
+    confirmButtonText: 'Sim, excluir!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      fetch("http://127.0.0.1:3000/veiculos/" + id, {  
+        method: "DELETE",
+      })
+      .then(() => {
+        listar();
+        Swal.fire('Excluído!', 'O veículo foi removido.', 'success');
+      })
+      .catch(err => console.error("Erro ao excluir veículo:", err));
+    }
+  });
 }
 
 function salvar() {
@@ -42,10 +57,10 @@ function salvar() {
 
   let url, metodo;
   if (idveiculoatual > 0) {
-    url = "http://127.0.0.1:3333/veiculo/" + idveiculoatual;
+    url = "http://127.0.0.1:3000/veiculos/" + idveiculoatual;
     metodo = "PUT";
   } else {
-    url = "http://127.0.0.1:3333/veiculo";
+    url = "http://127.0.0.1:3000/veiculos";
     metodo = "POST";
   }
 
@@ -77,7 +92,7 @@ function listar() {
   const listar = document.getElementById("lista");
   listar.innerHTML = "<tr><td colspan='5'>Carregando...</td></tr>";  
 
-  fetch("http://127.0.0.1:3333/veiculos")
+  fetch("http://127.0.0.1:3000/veiculos")
     .then(resp => resp.json())
     .then(dados => mostrar(dados))
     .catch(err => {
@@ -103,8 +118,8 @@ function mostrar(dados) {
       + "<td>" + dados[i].cor + "</td>"
       + "<td>" + precoFormatado + "</td>"
       + "<td>" 
-      + "<button type='button' class='btn btn-primary btn-sm' onclick='alterar(" + dados[i].id + ")'>Alterar</button> "
-      + "<button type='button' class='btn btn-secondary btn-sm' onclick='excluir(" + dados[i].id + ")'>Excluir</button>"
+      + "<button type='button' class='btn btn-editar btn-sm' onclick='alterar(" + dados[i].id + ")' title='Editar'><i class='bi bi-pencil-fill'></i></button> "
+      + "<button type='button' class='btn btn-excluir btn-sm' onclick='excluir(" + dados[i].id + ")' title='Excluir'><i class='bi bi-trash-fill'></i></button>"
       + "</td>"
       + "</tr>";
   }
